Modify sequential numbers mixin
