+++ global/data/initial-data.json
@@ -1,5 +1,5 @@
 {
-    "_migration_index": 20,
+    "_migration_index": 11,
     "organization": {
         "1": {
             "id": 1,
