+++ openslides_backend/action/actions/motion_block/create_update_delete.py
@@ -1,5 +1,3 @@
-from typing import Any, Dict
-
 from ....models.models import MotionBlock
 from ....permissions.permissions import Permissions
 from ...action_set import ActionSet
@@ -25,13 +23,6 @@ class MotionBlockCreate(
     CreateActionWithListOfSpeakersMixin,
 ):
     dependencies = [AgendaItemCreate, ListOfSpeakersCreate]
-
-    def update_instance(self, instance: Dict[str, Any]) -> Dict[str, Any]:
-        instance = super().update_instance(instance)
-        instance["sequential_number"] = self.get_sequential_number(
-            instance["meeting_id"]
-        )
-        return instance
 
 
 @register_action_set("motion_block")
