+++ openslides_backend/action/actions/motion_statute_paragraph/create_update_delete.py
@@ -1,20 +1,14 @@
-from typing import Any, Dict
-
 from ....models.models import MotionStatuteParagraph
 from ....permissions.permissions import Permissions
 from ...action_set import ActionSet
+from ...generics.create import CreateAction
 from ...mixins.sequential_numbers_mixin import SequentialNumbersMixin
 from ...util.default_schema import DefaultSchema
 from ...util.register import register_action_set
 
 
-class MotionStatuteParagraphCreate(SequentialNumbersMixin):
-    def update_instance(self, instance: Dict[str, Any]) -> Dict[str, Any]:
-        instance = super().update_instance(instance)
-        instance["sequential_number"] = self.get_sequential_number(
-            instance["meeting_id"]
-        )
-        return instance
+class MotionStatuteParagraphCreate(SequentialNumbersMixin, CreateAction):
+    pass
 
 
 @register_action_set("motion_statute_paragraph")
