+++ openslides_backend/action/actions/projector/create.py
@@ -1,7 +1,6 @@
-from typing import Any, Dict
-
 from ....models.models import Projector
 from ....permissions.permissions import Permissions
+from ...generics.create import CreateAction
 from ...mixins.sequential_numbers_mixin import SequentialNumbersMixin
 from ...util.default_schema import DefaultSchema
 from ...util.register import register_action
@@ -9,7 +8,7 @@ from ..meeting.shared_meeting import used_as_default_for_schema
 
 
 @register_action("projector.create")
-class ProjectorCreateAction(SequentialNumbersMixin):
+class ProjectorCreateAction(SequentialNumbersMixin, CreateAction):
     """
     Action to create a projector.
     """
@@ -39,10 +38,3 @@ class ProjectorCreateAction(SequentialNumbersMixin):
         },
     )
     permission = Permissions.Projector.CAN_MANAGE
-
-    def update_instance(self, instance: Dict[str, Any]) -> Dict[str, Any]:
-        instance = super().update_instance(instance)
-        instance["sequential_number"] = self.get_sequential_number(
-            instance["meeting_id"]
-        )
-        return instance
